+++ __tests__/index.test.ts
@@ -32,7 +32,6 @@ const runMock = jest.spyOn(main, 'run').mockImplementation();
 
 describe('index', () => {
   it('calls run when imported', async () => {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
     require('../src/index');
 
     expect(runMock).toHaveBeenCalled();